backend/requirements.txt
aniso8601==7.0.0
asgiref==3.4.1
Django==3.2.7
django-cors-headers==3.8.0
django-filter==2.4.0
django-graphql-jwt==0.3.4
graphene==2.1.9
graphene-django==2.15.0
graphene-django-pagination @ https://github.com/instruct-br/graphene-django-pagination/archive/master.zip
graphene-file-upload==1.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
Pillow==8.3.2
promise==2.3
PyJWT==2.1.0
pytz==2021.1
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
yapf==0.31.0
